Ignore loganalyzer error for MSN2700 running test_po_update.py
There's been repeated instances of loganalyzer failing for one or both
test cases in test_po_update.py. This is largely because both the port
channel member and port channel deletions get queued up in orchagent
before orchagent has a chance to process them, and because of the redis
DB priorities in orchagent, the port channel deletion gets processed
before the port channel member deletion.

From my understanding, the CPU on the 2700 is on the weaker side, so
this is going to happen frequently unless there's more of a delay
between the member deletion and the port channel deletion. As an easy
fix, ignore that error message for this platform.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>